update: change the lambda package
